Support for invite URLs and fixed game draws.
Server was not handling the case of a full board correctly and now it
has been fixed with some changes in the protocol (no_moves -> draw).
Added unit tests for game draw.

Added support for requesting private games and for users to join them.
Now the email a friend feature is working!